Remove UI state module, move WRTC stateful actions to WRTC page namespace
