+++ src/app/app.module.ts
@@ -31,7 +31,6 @@ import { DataStoreProvider } from '../providers/data-store/data-store';
 import { SecureStorage } from '@ionic-native/secure-storage';
 import { TestsModule } from '../modules/tests/tests.module';
 import { DeviceAuthenticationProvider } from '../providers/device-authentication/device-authentication';
-import { UiStateModule } from '../modules/ui-state/ui-state.module';
 
 @NgModule({
   declarations: [App],
@@ -45,7 +44,6 @@ import { UiStateModule } from '../modules/ui-state/ui-state.module';
     AppInfoModule,
     LogsModule,
     TestsModule,
-    UiStateModule,
   ],
   bootstrap: [IonicApp],
   entryComponents: [App],
