+++ src/pages/waiting-room-to-car/waiting-room-to-car.actions.ts
@@ -1,10 +1,34 @@
 import { Action } from '@ngrx/store';
 
 export const WAITING_ROOM_TO_CAR_VIEW_DID_ENTER = '[WaitingRoomToCarPage] Waiting roomn to car did enter';
+export const EYESIGHT_PASS_PRESSED = '[WaitingRoomToCar] Eyesight pass pressed';
+export const EYESIGHT_FAIL_PRESSED = '[WaitingRoomToCar] Eyesight fail pressed';
+export const EYESIGHT_FAIL_CANCEL_PRESSED = '[WaitingRoomToCar] Eyesight fail cancel pressed';
+export const EYESIGHT_FAIL_CONFIRMED = '[WaitingRoomToCar] Eyesight fail confirmed';
 
 export class WaitingRoomToCarViewDidEnter implements Action {
   readonly type = WAITING_ROOM_TO_CAR_VIEW_DID_ENTER;
 }
 
+export class EyesightPassPressed implements Action {
+  readonly type = EYESIGHT_PASS_PRESSED;
+}
+
+export class EyesightFailPressed implements Action {
+  readonly type = EYESIGHT_FAIL_PRESSED;
+}
+
+export class EyesightFailCancelPressed implements Action {
+  readonly type = EYESIGHT_FAIL_CANCEL_PRESSED;
+}
+
+export class EyesightFailConfirmed implements Action {
+  readonly type = EYESIGHT_FAIL_CONFIRMED;
+}
+
 export type Types =
-  | WaitingRoomToCarViewDidEnter;
+  | WaitingRoomToCarViewDidEnter
+  | EyesightPassPressed
+  | EyesightFailPressed
+  | EyesightFailCancelPressed
+  | EyesightFailConfirmed;
